Remove Progress package, add pip dependency instead
