--- deepfly/pose2d/progress/.gitignore
@@ -1,4 +1,0 @@
-*.pyc
-*.egg-info
-build/
-dist/
