+++ setup.py
@@ -24,6 +24,7 @@ setup(
         "torchvision",
         "opencv-python>=3.4.0.12",
         "tqdm",
-        "colorama"
+        "colorama",
+        "progress"
     ],
 )
